DFA creation, Buffer template and RegExp
Created DFA.ts for future DFA classes to inherit from. Some future DFAs will reference one type from the 8 types in TypeDef.ts. Each will depend on the function of the InputBuffer class to read the string

Added the Keywords.ts file to be able to compare identifiers with Keywords.